AST Cache checked more regularly. Bug fixes with building fuzzers
